pyproject.toml
[project]
name = "vaskup"
version = "0.1.0"
description = "VASKUP: Advanced Agentic RAG system for patent analysis using LangGraph"
authors = [
    {name = "VASKUP Team"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "streamlit>=1.28.0",
    "pandas>=2.0.0",
    "langchain>=0.1.0",
    "langgraph>=0.0.40",
    "langchain-community>=0.0.20",
    "langchain-openai>=0.0.8",
    "chromadb>=0.4.0",
    "requests>=2.31.0",
    "beautifulsoup4>=4.12.0",
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
    "openpyxl>=3.1.0",
    "lxml>=4.9.0",
    "jinja2>=3.1.0",
    "plotly>=5.15.0",
    "aiohttp>=3.8.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.5.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.black]
line-length = 88
target-version = ['py311']

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
